Pin period arrays to int32 in initial-conditions validation
`_collect_feasibility_errors` and `_check_regime_feasibility` built
`period` arrays via `jnp.array([...])` over Python ints, yielding
int64 under `jax_enable_x64`. Under the canonical-int32 invariant
these must be pinned to int32, otherwise an int64 `period` leaks into
the feasibility check and into user transition/leaf functions whose
`period` parameter is typed `int32`.

+++ src/lcm/simulation/initial_conditions.py
@@ -426,7 +426,9 @@ def _collect_structural_errors(
     else:
         # Validate that each subject's initial regime is active at their starting age.
         # Only safe to run when all ages are valid (so age_to_period lookup succeeds).
-        periods = jnp.array([ages.age_to_period(a.item()) for a in age_values])
+        periods = jnp.array(
+            [ages.age_to_period(a.item()) for a in age_values], dtype=jnp.int32
+        )
 
         active_mask = jnp.ones(regime_id_arr.size, dtype=bool)
         for regime_name, internal_regime in internal_regimes.items():
@@ -703,7 +705,8 @@ def _check_regime_feasibility(  # noqa: C901
         subject_states["age"] = initial_states["age"][idx_arr]
     if needs_period:
         subject_states["period"] = jnp.array(
-            [ages.age_to_period(a.item()) for a in initial_states["age"][idx_arr]]
+            [ages.age_to_period(a.item()) for a in initial_states["age"][idx_arr]],
+            dtype=jnp.int32,
         )
 
     # Split actions and params — actions are vmapped over, params are not
